Updated the project, formated dataset inputs and tokenised the test data
Added in LoRa model from transformers and set it up. yet to be trained

This an updated and working version of the classification process

Completed the fine tune model. Now going to add a PEFT model to the mix

PEFT.PY contains the Trained model with a PEFT wrapper to enhance efficiency